qa-tests: de-schedule tip-tracking (#16205)
The tip-tracking test has been superseded by the sync-from-scratch test
that is a sync from scratch plus a tip tracking, running erigon for 2
hours on the tip.
The tip-tracking test requires a pre-built database, and it is difficult
to ensure the correct database while there are many backward
incompatible changes on the main and release branches (MDBx versions,
new tables, different snapshot formats, etc.).
We will therefore transform the tip-tracking test into a "migrate +
tip-tracking" to test the migration procedure from Release 3.0 to
Release 3.1.
This PR remove the test scheduling from the main branch.