[r31] qa-tests: add migration procedure to the tip-tracking test (#16202)
This PR address 2 tip-tracking issues:

1) The new test scheduling policy, which is made up of a combination of
two labels ('chain' and 'test-type'), failed to address the tip-tracking
tests. This PR fixes this issue.

2) The tip-tracking test has been superseded by the sync-from-scratch
test that is a sync from scratch plus a tip tracking, running erigon for
2 hours on the tip.
The tip-tracking test requires a pre-built database, and it is difficult
to ensure the correct database while there are many backward
incompatible changes on the main and release branches (MDBx versions,
new tables, different snapshot formats, etc.).
We will therefore transform the tip-tracking test into a "migrate +
tip-tracking" to test the migration procedure from Release 3.0 to
Release 3.1.
This PR add the migration procedure to the tip-tracking test.