Update erigon-lib interface generation for Downloader changes (#16272)
In #15043 I didn't include the updated protobuf interfaces. I spotted
them in https://github.com/erigontech/erigon/pull/16081.

The vendoring is too brittle and makes it hard to generate local changes
in a workspace. I switched it to using a git submodule which is much
easier to work with. The interfaces generated here are direct outputs of
the protobuf so there's no need to include a reference to the interfaces
revision used in the Go code.