rpcdaemon: Add creation method on trace_ (#16518)
The creationMethod field is a specific feature of Geth's flatCallTracer,
which is invoked via the debug_ API. This tracer is not available in
Erigon's debug APIs. If the functionality provided by flatCallTracer is
considered valuable, a full implementation would be required within
Erigon (see eth/tracers/native/...)

Geth does not support tracing APIs prefixed with trace_. Instead, it
uses APIs prefixed with debug_.

Reth has added the creationMethod field to the output of its trace_
APIs.

This PR add creation method on tracing api and supersedes and replaces
https://github.com/erigontech/erigon/pull/12557, which can now be closed