[r3.1] qa-tests: improve tip-tracking timeout and error reporting  (#16686)
The latest "tip-tracking & migration" executions show that it takes more
than ten hours to run two erigon cycles, so this PR raises the timeout
to more than 16 hours (8 hours are expected per cycle).
Furthermore, when the test is cancelled due to a timeout, GitHub
displays a grey icon, which is misleading. This PR should change it to
red, forcing a check to be performed. The error condition deliberately
treats false positives as errors, i.e. the test not being performed on
erigon during upgrades and downgrades.