polygon: Use BorSpansIndex to calculate span id at block number (#16683)
This PR changes the way the span id at a given block number is
calculated. Whereas before, the span id was calculated via the
`SpanIdAt(blockNum)` function, which did a simple arithmetic calculation
to get the span id, with the changes in this PR a `spanRangeIndex`
`RangeIndexer` is used to calculate the span at given block via a
`span.StartBlock -> (span.Id, span.EndBlock)` MDBX table
(`kv.BorSpansIndex` for `Span`s, and `kv.BorProducerSelectionsIndex` for
`BorProducerSelections` ).

The span id for a given `blockNum` is now calculated as:
`max span.Id such that blockNum >= span.StartBlock && blockNum <=
span.EndBlock`

The reasons for this change are 2:

1. In the upcoming VeBlop hardfork
(https://github.com/maticnetwork/Polygon-Improvement-Proposals/blob/main/PIPs/PIP-64.md)
spans could overlap if there is a producer rotation before the current
span completes. E.g. Span 0 is [0 ; 6399], but in block 5 there is a new
span with a different producer (due to the block producer of span 0
going down), so Span 1 is [5; 12,799] . Therefore, the span at block 10
for example is no longer Span 0, but Span 1 instead. For this reason, a
more flexible approach is needed to determine the span id for a given
block.

2. In the old code a hardcoded span length of 6400 blocks was used. This
worked fine for Bor mainnet and Amoy, but for Kurtosis testing, Polygon
uses a configurable span length (default now is 128 blocks). This means
that the Erigon node in the kurtosis setup was misbehaving. With this
PR, the kurtosis Polygon tests can work no matter what the span length
is, since the span id calculation is more flexible.

The old `SpanIdAt()` function is deprecated and only kept in some
obsolete tests that use `HeimdallSimulator`, which can be removed in the
future.

---------

Co-authored-by: antonis19 <antonis19@users.noreply.github.com>