integration: execute run_migrations also on consensus db (#16711)
Starting from #16678, we've tried using `integration run_migrations`
within RPC Integration Tests CI workflows to handle creation of new
tables automatically: it worked immediately for Ethereum, but failed for
both Gnosis and Polygon.

Some changes to `integration run_migrations` are necessary to make it
work for them:

- Gnosis: migrations must work also for `consensus` database, which is
used there because pre-Merge consensus was AuRa. `ConsensusTables` used
within `consensus` database already contained all `ChaindataTables`, so
just running the migrations on the `consensus` database is sufficient.
- Polygon: migrations must work also for both `heimdall` and
`polygon-bridge` databases, which are used by internal Polygon
components. In order to use exactly the same run migration procedure, we
need to create all `ChaindataTables` also there. Of course, this is not
ideal but IMO should be addressed as a separate issue (namely,
refactoring and improvement of run migration procedure not to require
the same tables in all databases).