VeBlop: Update span index lookup and add more unit tests (#16729)
Follow-up to : https://github.com/erigontech/erigon/pull/16683

The previous PR relied on the assumption that `span.StartBlock` is
monotonically increasing, but it turns out that assumption is not true
due to a future span being announced in advance from Heimdall and a span
rotation taking place before the announced span.

e.g.

```Go
var spans = []Span{
		Span{
			Id:         0,
			StartBlock: 0,
			EndBlock:   999,
		},
		Span{
			Id:         1, // new span announced
			StartBlock: 1000,
			EndBlock:   1999,
		},
		Span{
			Id:         2, // span rotation
			StartBlock: 5,
			EndBlock:   1999,
		},
	}
```

Therefore, this PR adapts the lookup algorithm by starting the search
from `entry := cursor.Seek(blockNum)` and walking the `BorSpans` table
backwards using `cursor.Prev()` while recording the highest span id
which is in range, until an out-of-range span is reached, at which point
we break and return the highest span id seen.

This lookup algorithm relies on the assumption that there won't ever be
more than 2 consecutive spans for which `span1.Id < span2.Id` and
`span1.StartBlock > span2.StartBlock` . If this assumption were ever to
be broken, we would have to fall back on brute-force linear search.

Also, another detail to take into account is that due to the span
rotations multiple spans with the same `StartBlock` could be produced,
therefore overwriting the value of the index table at that block number.
This is not a problem, because if this happens, then the previous span
value will become entirely obsolete and no block number could have the
old span associated with it.

---------

Co-authored-by: antonis19 <antonis19@users.noreply.github.com>