integration: fix database not initialized in run_migrations (#16762)
After adding `integration run_migrations` step within RPC Integration
Tests, some runs started to fail intermittently on Gnosis and Polygon
with error "database is not initialized":


https://github.com/erigontech/erigon/actions/runs/17097843879/job/48486419048

https://github.com/erigontech/erigon/actions/runs/17097843833/job/48486419046

https://github.com/erigontech/erigon/actions/runs/17099664503/job/48492523550

It turns out the unpredictable order of range iteration over `dbPaths`
map can lead any database different from `chaindata` to be opened as
first, which triggers the `openSnapshotOnce` function to be executed on
a database with empty `Config` table, hence the error.

This guarantees that `chaindata` is always migrated as first.