polygon: Implement VeBlop consensus changes (#16797)
This PR incorporates the VeBlop consensus changes to Astrid.

The current PR, together with
https://github.com/erigontech/erigon/pull/16739 implement Polygon's
VeBlop hard fork on Erigon.

The main change is related to waiting for a new span (called span
rotation), in case the VeBlop producer is unavailable or misbehaving.
This time-limited synchronization step is needed to process the new span
that will update the current producer, so that block production can
continue.

---------

Co-authored-by: antonis19 <antonis19@users.noreply.github.com>