fix rebuilding commitment by smaller shards (#16779)
so file of .0-1024.kv is representing `range` of steps [0, 1024) 
we do split amount of keys in accounts and storage evenly across several
`shards` with cap to 128 steps per shard.

This shard is part of future file of 0-1024.kv. it does not represent
latest values as of step 128. It is just a dump of commitment branches
representing trie as of step 1024.

---------

Co-authored-by: alex <AskAlexSharov@gmail.com>