VeBlop: Skip span sorting by StartBlock in EntityFetcher (#16815)
Due the Veblop hard fork the span sorting by `span.StartBlock` does not
make sense because the `StartBlock` will not be strictly increasing. The
sorting by `StartBlock` resulted in a span id gap in `ObserveSpan()` due
to a span with a lower Id but higher StartBlock being relocated to
higher indexes in the sorted array of fetched spans.

Therefore, this PR disables the sorting, and just relies on the
`span.Id` for the order.

---------

Co-authored-by: antonis19 <antonis19@users.noreply.github.com>