execution/engineapi: enable bbd v2 (#16673)
closes https://github.com/erigontech/erigon/issues/16424
closes https://github.com/erigontech/erigon-qa/issues/74

This PR enables the new EL backward downloader flow by default. It was
introduced in a previous PR
https://github.com/erigontech/erigon/pull/16270 and has undergone
testing
https://github.com/erigontech/erigon/pull/16673#issuecomment-3241939872.

Few issues were found in the new flow and fixed during testing:
- fixes a nil pointer when marking a bad block
- the new backward block downloader was using `block.HashCheck()` for
making sure the peer responded with the correct body for the given
header, however HashCheck is doing some additional validation checks on
top of just matching the Transaction, Uncle, Withdrawal hashes to the
header hashes - to solve this I added a `body.MatchesHeader(h)` func
just for these 3 which HashCheck then re-uses. This allows the backward
downloader to just focus on downloading the blocks and allows the
additional validation checks done in `HashCheck` to be performed later
on in higher abstractions
- fixes a race condition on startup that was happening in hive tests due
to peer connection events - solution was to add logic to replay
currently connected good peers when hitting the Sentry "PeerEvents"
subscription API