execution: more accurate bad block responses (#16994)
fixes https://github.com/erigontech/erigon/issues/16973

The problem is that in the fork validator we always return
`consensus.ErrInvalidBlock` when running the sync loop via `StateStep`
fails:
```
if err := stages2.StateStep(ctx, chainReader, backend.engine, txc, stateSync, header, body, unwindPoint, headersChain, bodiesChain, config.ImportMode); err != nil {
	logger.Warn("Could not validate block", "err", err)
	return errors.Join(consensus.ErrInvalidBlock, err)
}
```

We already correctly mark the blocks as invalid (when they are indeed
invalid) and always unwind with a `BadBlock` UnwindReason in those
situations.

So we can make use of that and simply pass up the
`consensus.ErrInvalidBlock` from the `BadBlock` UnwindReason up to the
caller of the sync loop `Run` and `RunNoInterupt` functions. This way
when we call Run from a fork choice update it can return
`consensus.ErrInvalidBlock` correctly. Note that for the fork choice
update code path we do not return an err from the sync loop run when we
hit a bad block (we only unwind with a bad block reason) - because we
run it with `badBlockHalt: false`.

For the `StateStep` flow - it runs the execution loop with
`badBlockHalt: true`. That code path always returns the err so we need
to make sure that all relevant error paths return `ErrInvalidBlock` when
they should - then we can get rid of the the logic that always does
`errors.Join(consensus.ErrInvalidBlock, err)` return for every error
from `StateStep`.

Other parts (inside `StateStep`) that return `consensus.ErrInvalidBlock`
are:
- errors from `engine.VerifyHeader` (in `addAndVerifyBlockStep`) - as
expected
- errors from `engine.VerifyUncles` (in `addAndVerifyBlockStep`) - as
expected

All other errors returned are operational errors and not
`consensus.ErrInvalidBlock` errors.