Fixed merge snapshots range selection algorithm (#17032)
The gaps may have occurred inside the snapshot files, partly because the
mechanism for calculating merge intervals did not take into account that
snapshots can have different block heights. For example, in this case
the merge algorithm would be triggered even though one of the snapshot
files is completely missing:
```
0009-0010.borspans
0009-0010.borevents
0008-0009.checkpoints
```

I added an align check that should account for missing intervals for
certain types of snapshots.