tests: extract mock cl and engine api tester from shutter (#17103)
this is preparation for following PRs which add engine api tests (in
particular for reorg scenarios - relates to
https://github.com/erigontech/erigon/issues/17025)

it extracts the MockCl from the shutter block building integration test
and generalises it a bit more so it is re-usable in more generic engine
api test scenarios (e.g. used in
https://github.com/erigontech/erigon/pull/17105)

it also extracts the initialisation logic into a new EngineApiTester
which can also be re-used for different generic engine api test
scenarios (e.g. it can be used for the new "blockchain engine x test
format" - https://github.com/erigontech/erigon/issues/16562)