Fix incorrect `GoodPeers` count (#17171)
After eth/69, we were calling `getOrCreatePeer` before the handshake,
however in the case of a failed handshake these peers were not being
cleaned up - leading to incorrect peer counts in the `GoodPeers` log.

Since there is a constraint of setting up `defer`s in the correct order,
we create the `peerInfo` after the handshake and store the handshake
data in temporary variables.