Fix parameter expansion in shell code. (#17224)
Fix parameter expansion (introduced in #17221 and #17212) in shell code
which could lead to broken artifact file names, etc.

As of this change:
```
-        DOCKER_TAGS: ${{ env.DOCKERHUB_REPOSITORY }}:$RELEASE_VERSION
+        DOCKER_TAGS: "${{ env.DOCKERHUB_REPOSITORY }}:${{ inputs.release_version }}"
```
You can't define an environment variable using another environment
variable in the same env block of a GitHub Actions workflow. This is a
known limitation of how GitHub Actions processes variables. The core
reason is that the env block is processed before the step runs, and
variables are not evaluated in a way that allows for self-referential or
dependent definitions within the same block.

(cherry picked from commit e329b8b14d2dcbc4b5d8b75ed4583a9e045185c6)