rpcdaemon: debug_accountRange support ALSO interface as GETH (#17146)
This PR introduces few changes to the debug_accountRange() method to
improve its compatibility with Geth. The changes are as follows:

* start Parameter: The start parameter now supports both the []byte
format (used currently by Erigon) and the hexutil.Bytes format (as a
hexadecimal string, used by Geth). The []byte format is now considered
deprecated. We can can remove in next versions.

* incompletes Parameter: Added the optional incompletes parameter to the
API for Geth compatibility. By default, its value is false. It's
important to note that setting incompletes to true is not yet supported,
as this functionality is specific to Geth's implementation.

RPC-tests add tests according Geth interface