Use MaxUint64 as return value to `HistoryStartFrom()` if no files present (#17308)
Currently default return value of 0 leads to `PrunedError` not being
thrown when commitment history is not available when `eth_getProof` is
invoked. This leads to incorrect root hash error, but instead a
`PrunedError` should be thrown.

A default value of +∞ is more appropriate, and fixes
https://github.com/erigontech/erigon/issues/16776 as well.

UPDATE: It could happen that there are no history files, but there is
history in MDBX, so the function is now updated to look into MDBX for
the first txnum if there are history files in snapshots.

---------

Co-authored-by: antonis19 <antonis19@users.noreply.github.com>