Split exec3 inner functions into  exec3_serial & exec3_parallel (#17338)
This refactor moved the parallel and serial implementations in exec3.go
into thier respective source files exec3_serial.go and
exec3_parallel.go.

Although mainly an organizational change this also clarifies the meaning
and use of a few captive variables which where used both in the inner
closures and the outer scope of the bounding function.