Make access list trace order deterministic (#17291)
This is a fix to avoid access list tracing from returning results in
random map order by adding code to preserve insetion order.

This will make access list RPC's return the same results when called
with the same input. I made the fix to address what I thought was a
regression - it seems with the right version of json-diff this is not
strictly necessary - however I think its a reasonable fix to include.