Removed manual flush in json stream buffer (#17382)
This PR removes manual stream.Flush in the code, where it is not
necessary. The idea is Flush should not be called in the code which does
not control io.Writer. This is the responsibility of the caller function
to ensure that Flush is called if it wants to use io.Writer content.