Make loop exhausted explicit (#17455)
This fixes two issues when deleting chaindata where the last frozen step
ends imid block.

1. Error handling at the end of serial processing is adjusted to
explicitly handle `ErrLoopExhausted` so that underlying error's are not
masked by the final frozen commitment error.
2. It fixes a subseuent state mismatch error due to the end of block
transactions not being read with the historical reader and hence
presenting the latest state - which has been updated in the frozen
blocks. This happens when the `state` key is not stored for the final
transaction in the block.

The second fix means that processing continues but leaves open the
following questions which need further investigation:

1. Why is the final cimmitment state not stored at the last tx of the
step (in the case here it was 9998 instead of 9999, so ine short of the
end of the step.
2. What happens to the written output of the executed transactions. I
assume at the moment that these are persisted in the DB, which leaves a
duplicate. I think this is an error, though harmless assuming the
exeution produces the same result.