Unwrap Shared Doman and Temporal Tx Params (#17402)
This PR changes the way SharedDomains and TemporalTx's are passed to
stage components. Rather than wraping them inside of a TxContainer.

The reason for doing this is that the container removes type and content
saftey when we're passing domains an transactions. For the most part now
the tx passed is temporal - and all of the underlying code now handles
temproal transactions, so I think for the stage loop thie wrapper has
become redundant.

Why Now ? Becuase I have discovered an issue with unwinds when running
parallel processing with an external tx. This is becuase in the case the
shared domain is not shared, or its memory gets clear in unwind, and
becuase the unwind flush is to the external RoTx of the loop, the RoTx's
of the parallel workers can't see this update as its not been commited
yet.

To fix this I need to trace the SD usage and use an SD which is shared
between unwind and exec. This will make that easier to do.