Fix historical state queries from commitment (#17413)
Related to second example in
https://github.com/erigontech/erigon/issues/16752

When reading certain accounts or storage slots from history, they may
not have been created yet as of that block/txnum. Then the return value
should be nil, and not found. But in the code there is extra logic to
fall back to reading the latest version of the account from files. This
is incorrect when querying historical values, but it is needed when
commitment is rebuilt on existing domain files (which would have
`withHistory=false`).

This PR disables this fallback behavior when `withHistory=true`, so that
the accurate state values may be read when generating historical merkle
proofs.

---------

Co-authored-by: antonis19 <antonis19@users.noreply.github.com>