rpcdaemon:  pre-byzantium tx calculate postState (#17479)
This PR fixes the following  Hive RPC compatibility tests:
- eth_getBlockReceipts/get-block-receipts-by-hash (erigon_default)
- eth_getBlockReceipts/get-block-receipts-n (erigon_default)

For pre-Byzantium blocks, the Ethereum specification requires that
transaction receipts contains:
-  the "postState" field and 
- "NOT" the `status` field.

To calculate "postState" it is necessary enable the commitment
The current database and snapshots for pre-Byzantium blocks do not
contain valid `postState` data, therefore, we must explicitly
re-calculate these receipts to be compliant to the specification
(`postState` instead of `status`).

If "commitment is not enabled", we always provide the "status" field,
regardless of the block number.

Some tests has been implemented to see the fixes using the hive DB

In order to see these tests successfully solved in Hive, a PR enabling
commitment is required in the Hive repository itself.